remove core package for now
